Fix dueDate Calculation
