+++ lib/ViewItem/ViewItem.js
@@ -2,14 +2,14 @@ import _ from 'lodash';
 import moment from 'moment'; // eslint-disable-line import/no-extraneous-dependencies
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormattedDate, FormattedTime } from 'react-intl';
+import { FormattedDate } from 'react-intl';
 import MultiColumnList from '@folio/stripes-components/lib/MultiColumnList';
 import Button from '@folio/stripes-components/lib/Button';
 import { UncontrolledDropdown } from '@folio/stripes-components/lib/Dropdown';
 import MenuItem from '@folio/stripes-components/lib/MenuItem';
 import DropdownMenu from '@folio/stripes-components/lib/DropdownMenu';
 
-import { getDueDate } from '../../util';
+import { formatTime } from '../../util';
 
 const sortMap = {
   'No.': loan => loan.no,
@@ -83,8 +83,8 @@ class ViewItem extends React.Component {
       Title: loan => _.get(loan, ['item', 'title']),
       'Loan Policy': loan => _.get(loan, ['loanPolicy', 'name']),
       barcode: loan => _.get(loan, ['item', 'barcode']),
-      'Due date': loan => (<FormattedDate value={getDueDate(loan)} />),
-      Time: loan => (<FormattedTime value={getDueDate(loan)} />),
+      'Due date': loan => (<FormattedDate value={loan.dueDate} />),
+      Time: loan => (formatTime(loan.dueDate)),
       ' ': loan => this.renderActions(loan),
     };
   }
