+++ util.js
@@ -1,10 +1,9 @@
 import { get } from 'lodash';
 import moment from 'moment'; // eslint-disable-line import/no-extraneous-dependencies
-import {
-  defaultPatronIdentifier,
-  patronIdentifierMap,
-  loanProfileTypesMap,
-} from './constants';
+import React from 'react';
+import { FormattedTime } from 'react-intl';
+
+import { defaultPatronIdentifier, patronIdentifierMap } from './constants';
 
 // serialized object into http params
 export function toParams(obj) {
@@ -17,34 +16,10 @@ export function getFullName(user) {
     ${get(user, ['personal', 'middleName'], '')}`;
 }
 
-export function isRollingProfileType(loanProfile) {
-  return (loanProfile.profileId === loanProfileTypesMap.ROLLING ||
-    loanProfile.profileId === 'ROLLING');
-}
-
-export function isFixedProfileType(loanProfile) {
-  return (loanProfile.profileId === loanProfileTypesMap.FIXED ||
-    loanProfile.profileId === 'FIXED');
-}
-
-export function getDueDate(loan) {
-  const loanPolicy = loan.loanPolicy;
-  const loanProfile = loanPolicy.loansPolicy || {};
-
-  // fixed type
-  if (isFixedProfileType(loanProfile) && loanPolicy.loanable) {
-    if (loanPolicy.fixedDueDateSchedule) {
-      return loanPolicy.fixedDueDateSchedule.schedule.due;
-    }
-  }
-
-  return loan.dueDate;
-}
-
-export function getFixedDueDateSchedule(schedules) {
-  const today = moment(new Date());
-  return schedules.find(s =>
-    today.isBetween(moment(s.from).startOf('day'), moment(s.to).endOf('day')));
+export function formatTime(dateStr) {
+  if (!dateStr) return dateStr;
+  const localDateStr = moment(dateStr).local().format();
+  return (<FormattedTime value={localDateStr} />);
 }
 
 export function getPatronIdentifiers(checkoutSettings) {
